+++ Oni/Resources/Tags/ABNA.h
@@ -33,7 +33,7 @@ public:
 	~ABNA();
 	void Load(char *data);
 	int32_t* GetInstanceIDs();
-	uint64_t DataSize();
+	int32_t DataSize();
 	char* ExportData();
 };
 #endif
