+++ Oni/Resources/Tags/AGQC.h
@@ -35,7 +35,7 @@ public:
 	~AGQC();
 	void Load(char *data);
 	int32_t* GetInstanceIDs();
-	uint64_t DataSize();
+	int32_t DataSize();
 	char* ExportData();
 };
 #endif
