+++ Oni/Resources/Tags/AGQG.h
@@ -44,7 +44,7 @@ public:
 	~AGQG();
 	void Load(char *data);
 	int32_t* GetInstanceIDs();
-	uint64_t DataSize();
+	int32_t DataSize();
 	char* ExportData();
 };
 #endif
